+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/tsv/impl/SSMMetadataTSVConverter.java
@@ -1,8 +1,9 @@
-package org.icgc.dcc.pcawg.client.tsv.converters;
+package org.icgc.dcc.pcawg.client.tsv.impl;
 
 import lombok.NoArgsConstructor;
 import org.icgc.dcc.pcawg.client.model.ssm.metadata.SSMMetadata;
 import org.icgc.dcc.pcawg.client.model.ssm.metadata.SSMMetadataFieldMapping;
+import org.icgc.dcc.pcawg.client.tsv.TSVConverter;
 
 import static org.icgc.dcc.common.core.util.stream.Streams.stream;
