+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/tsv/impl/SSMPrimaryTSVConverter.java
@@ -1,8 +1,9 @@
-package org.icgc.dcc.pcawg.client.tsv.converters;
+package org.icgc.dcc.pcawg.client.tsv.impl;
 
 import lombok.NoArgsConstructor;
 import org.icgc.dcc.pcawg.client.model.ssm.primary.SSMPrimary;
 import org.icgc.dcc.pcawg.client.model.ssm.primary.SSMPrimaryFieldMapping;
+import org.icgc.dcc.pcawg.client.tsv.TSVConverter;
 
 import static org.icgc.dcc.common.core.util.stream.Streams.stream;
